Set the default alpha value to 0
The problem:
When using Compton/Picom, one can not use xsetroot to change the
background and hsetroot is needed (more details here:
https://github.com/chjj/compton/issues/162). However, when setting the
background with hsetroot, even in the very top of .xinitrc, there still
is a short blink of gray on fast-loading systems. This behavior does not
occur with xsetroot, as the initial/default color is black.

The solution in this commit:
Set the default alpha value from 255 to 0, so that the initial/default
color will be black as well, thus mimicing the behavior of xsetroot.

Signed-off-by: Leonid Bloch <lb.workbox@gmail.com>